separate benchmark and pipelined benchmark
