+++ src/tools/nbench/benchmark_pipelined.hpp
@@ -24,9 +24,10 @@
 #include "ngraph/function.hpp"
 #include "ngraph/runtime/performance_counter.hpp"
 
-std::vector<ngraph::runtime::PerformanceCounter> run_benchmark(std::shared_ptr<ngraph::Function> f,
-                                                               const std::string& backend_name,
-                                                               size_t iterations,
-                                                               bool timing_detail,
-                                                               int warmup_iterations,
-                                                               bool copy_data);
+std::vector<ngraph::runtime::PerformanceCounter>
+    run_benchmark_pipelined(std::shared_ptr<ngraph::Function> f,
+                                  const std::string& backend_name,
+                                  size_t iterations,
+                                  bool timing_detail,
+                                  int warmup_iterations,
+                                  bool copy_data);
